requirements.txt
celery==5.2.3
djangorestframework==3.13.1
django-celery-beat==2.2.1
requests==2.27.1
gevent==21.12.0
django-cors-headers==3.13.0
django-environ==0.9.0